linux: Open files with `O_CLOEXEC` to not leak fds to child processes
In general, `open()` should always be called with `O_CLOEXEC`, except
immediately before an `exec*` call that will use the fd.

Previously the fd was leaked to a child process, as seen here:

```c

int main() {
    assert(hid_open_path("/dev/hidraw0") > 0);
    char *args[] = {"/bin/sh", "-c", "readlink /proc/self/fd/*", NULL};
    execve("/bin/sh", args, NULL);
}
```

In some cases this could be a security issue if child processes may be
less privileged and shouldn't have access to the file.