Add pp_data_dump tool (#467)
Add a small command line tool for Windows, which dumps the content of the `Preparsed Data` structure, provided by the Windows HID subsystem, into a file.

The generated file is in a text format, which is readable for human, as by the `hid_report_reconstructor_test.exe` unit test executable of the HIDAPI project. This unit test allows it to test the HIDAPI's report descriptor reconstructor - offline, without the hardware device connected.