hid: add error data to hid_device structure
Store libusb error code so it can be retrieved later.

Includes the original error code as well as a context-specific message
which the libusb documentation sometimes specifies for each function.
Code which uses those functions are meant to set the contextual message
whenever possible.

The code is initialized to a success state which implies no errors yet.
The contextual error message is initialized to NULL and is not freed
when the device is closed. It is meant to point at string literals.