hid: define easy error data setter function
Sets all error data, including an optional contextual error message
which is supposed to be a non-freeable constant string
such as a string literal.

Contextual error messages are meant to be used in the cases
the libusb documentation goes into detail as to what happened.
Passing NULL will produce a message with just the libusb_error_name
and the libusb_strerror results. Passing a string literal will produce
a message that contains the additional context in addition to the error
name and message.